backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.0
bcrypt>=4.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
better-exceptions==0.3.3
psycopg2-binary>=2.9.0

# Phase III: AI Chatbot dependencies
openai>=1.0.0
tiktoken>=0.5.0